+++ app/controllers/StatusController.java
@@ -11,7 +11,7 @@ import play.mvc.Result;
 import javax.inject.Inject;
 import java.util.List;
 
-public class StatusController extends Controller
+public class StatusController extends ApplicationController
 {
     private JPAApi jpaApi;
     private FormFactory formFactory;
@@ -26,6 +26,8 @@ public class StatusController extends Controller
     @Transactional(readOnly = true)
     public Result getStatuses()
     {
+        if (isLoggedIn()&& getLoggedInSiteAdminRole().equals("Admin"))
+        {
         DynamicForm form = formFactory.form().bindFromRequest();
         String sql = "SELECT s FROM TicketStatus s " +
                 " WHERE statusName LIKE :searchCriteria " +
@@ -42,12 +44,18 @@ public class StatusController extends Controller
 
 
         return ok(views.html.Status.statusList.render(ticketStatuses, searchCriteria));
+        } else
+        {
+            return redirect(routes.AdministrationController.getLogin());
+        }
 
     }
 
     @Transactional(readOnly = true)
     public Result getStatus(Integer statusId)
     {
+        if (isLoggedIn()&& getLoggedInSiteAdminRole().equals("Admin"))
+        {
         String sql = "SELECT s FROM TicketStatus s " +
                 "WHERE statusId = :statusId";
         TicketStatus ticketStatus = jpaApi.em().createQuery(sql, TicketStatus.class).
@@ -55,11 +63,17 @@ public class StatusController extends Controller
 
 
         return ok(views.html.Status.status.render(ticketStatus));
+        } else
+        {
+            return redirect(routes.AdministrationController.getLogin());
+        }
     }
 
     @Transactional
     public Result postStatus(Integer statusId)
     {
+        if (isLoggedIn()&& getLoggedInSiteAdminRole().equals("Admin"))
+        {
         String sql = "SELECT s FROM TicketStatus s " +
                 "WHERE statusId = :statusId";
 
@@ -74,16 +88,28 @@ public class StatusController extends Controller
         jpaApi.em().persist(ticketStatus);
 
         return redirect(routes.StatusController.getStatuses());
+    } else
+    {
+        return redirect(routes.AdministrationController.getLogin());
+    }
     }
 
     public Result getNewStatus()
     {
+        if (isLoggedIn()&& getLoggedInSiteAdminRole().equals("Admin"))
+        {
         return ok(views.html.Status.newstatus.render());
+        } else
+        {
+            return redirect(routes.AdministrationController.getLogin());
+        }
     }
 
     @Transactional
     public Result postNewStatus()
     {
+        if (isLoggedIn()&& getLoggedInSiteAdminRole().equals("Admin"))
+        {
         DynamicForm form = formFactory.form().bindFromRequest();
         String status1 = form.get("ticketStatus");
         TicketStatus ticketStatus = new TicketStatus();
@@ -92,15 +118,25 @@ public class StatusController extends Controller
 
 
         return redirect(routes.StatusController.getStatuses());
+    } else
+    {
+        return redirect(routes.AdministrationController.getLogin());
+    }
     }
     @Transactional
     public Result deleteStatus(int statusId)
     {
+        if (isLoggedIn()&& getLoggedInSiteAdminRole().equals("Admin"))
+        {
         String sql = "SELECT ts FROM TicketStatus ts " +
                 "WHERE statusId = :statusId";
         TicketStatus ticketStatus = jpaApi.em().createQuery(sql, TicketStatus.class).
                 setParameter("statusId", statusId).getSingleResult();
         jpaApi.em().remove(ticketStatus);
         return redirect(routes.StatusController.getStatuses());
+        } else
+        {
+            return redirect(routes.AdministrationController.getLogin());
+        }
     }
 }
